pull out supplementary data updates
